Refactor and working example in main.rs
